requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq
faiss-cpu
pypdf
